___JmolVersion="13.1.15_dev_2013.04.21"
TODO: PyMOL movies, electron density, CGO

bug fix: PyMOL label fix, including fonts



Former-commit-id: 9e9aa144ae91e057bacd2200ea954348efb37ea8

+++ Jmol/src/org/jmol/export/_TachyonExporter.java
@@ -230,7 +230,7 @@ public class _TachyonExporter extends __RayTracerExporter {
   protected void fillConicalCylinder(P3 screenA, P3 screenB,
                                      int madBond, short colix, byte endcaps) {
     // conic sections not implemented in Tachyon
-    int diameter = viewer.scaleToScreen((int) ((screenA.z + screenB.z)/2f), madBond);
+    int diameter = (int) viewer.scaleToScreen((int) ((screenA.z + screenB.z)/2f), madBond);
     fillCylinderScreenMad(colix, endcaps, diameter, screenA, screenB);
    }
