+++ Jmol/src/org/jmol/renderbio/BackboneRenderer.java
@@ -62,7 +62,7 @@ public class BackboneRenderer extends BioShapeRenderer {
       if (mad < 0) {
         g3d.drawLine(colixA, colixB, xA, yA, zA, xB, yB, zB);
       } else {
-        int width = (exportType == GData.EXPORT_CARTESIAN ? mad : viewer
+        int width = (int) (exportType == GData.EXPORT_CARTESIAN ? mad : viewer
             .scaleToScreen((zA + zB) / 2, mad));
         g3d.fillCylinderXYZ(colixA, colixB, GData.ENDCAPS_SPHERICAL, width, xA,
             yA, zA, xB, yB, zB);
